refactored catalog interface and control writing
